看起来不像是代码问题,有点怀疑int转float了,回滚下训练试一下;

+++ SMG_NothingIsAll/AIFoundation/AINet/AIPort/AIPort.h
@@ -66,8 +66,8 @@
  */
 @interface AISPStrong : NSObject <NSCoding>
 
-@property (assign,nonatomic) CGFloat sStrong;
-@property (assign,nonatomic) CGFloat pStrong;
+@property (assign,nonatomic) NSInteger sStrong;
+@property (assign,nonatomic) NSInteger pStrong;
 
 @end
